+++ packages/ui/src/components/templates/report/components/BusinessReportSummary/BusinessReportSummary.tsx
@@ -1,60 +1,48 @@
 import React, { ComponentProps, FunctionComponent } from 'react';
-import { getSeverityFromRiskScore, Severity, SeverityType } from '@ballerine/common';
+import {
+  getSeverityFromRiskScore,
+  MERCHANT_REPORT_RISK_LEVELS_MAP,
+  MerchantReportRiskLevel,
+  RiskIndicatorRiskLevel,
+  Severity,
+  SeverityType,
+} from '@ballerine/common';
 import { ctw, severityToClassName } from '@/common';
 import { toTitleCase } from 'string-ts';
 import { Badge, Card, CardContent, CardHeader, RiskIndicatorsSummary } from '@/components';
 import { TextWithNAFallback } from '@/components/atoms/TextWithNAFallback';
+import { RiskIndicatorSchema } from '@ballerine/common';
+import { z } from 'zod';
 
 export const BusinessReportSummary: FunctionComponent<{
   summary: string;
   ongoingMonitoringSummary?: string;
-  riskLevels: {
-    legalRisk: SeverityType;
-    chargebackRisk: SeverityType;
-    reputationRisk: SeverityType;
-    transactionLaunderingRisk: SeverityType;
-  };
-  riskIndicators: Array<{
+  sections: ReadonlyArray<{
     title: string;
     search: string;
-    violations: Array<{
-      label: string;
-      severity: string;
-    }> | null;
+    riskIndicators: z.infer<typeof RiskIndicatorSchema>[] | null;
   }>;
-  riskScore: number;
+  riskLevel: MerchantReportRiskLevel;
   homepageScreenshotUrl: string | null;
   Link: ComponentProps<typeof RiskIndicatorsSummary>['Link'];
-}> = ({
-  riskIndicators,
-  summary,
-  ongoingMonitoringSummary,
-  riskLevels,
-  riskScore,
-  homepageScreenshotUrl,
-  Link,
-}) => {
-  const severity = getSeverityFromRiskScore(riskScore);
-
+}> = ({ sections, summary, ongoingMonitoringSummary, riskLevel, homepageScreenshotUrl, Link }) => {
   return (
     <div className={'grid grid-cols-5 gap-8'}>
       <Card className={!homepageScreenshotUrl ? 'col-span-full' : 'col-span-3'}>
         <CardHeader className={'pt-4 font-bold'}>
           <span className={'mb-1'}>Overall Risk Level</span>
           <div className="flex items-center space-x-2">
-            {(riskScore || riskScore === 0) && (
+            {riskLevel && (
               <Badge
                 className={ctw(
-                  severityToClassName[
-                    (severity?.toUpperCase() as keyof typeof severityToClassName) ?? 'DEFAULT'
-                  ],
+                  severityToClassName[riskLevel],
                   {
-                    'text-background': severity === Severity.CRITICAL,
+                    'text-background': riskLevel === MERCHANT_REPORT_RISK_LEVELS_MAP.critical,
                   },
                   'min-w-20 rounded-lg font-bold',
                 )}
               >
-                {toTitleCase(severity ?? '')} Risk
+                {toTitleCase(riskLevel)} Risk
               </Badge>
             )}
           </div>
@@ -106,7 +94,7 @@ export const BusinessReportSummary: FunctionComponent<{
         </Card>
       )}
 
-      <RiskIndicatorsSummary riskIndicators={riskIndicators} Link={Link} />
+      <RiskIndicatorsSummary sections={sections} Link={Link} />
     </div>
   );
 };